Improve robustness of cert retrieval
The [`sleep(5)`](https://github.com/ibmjstart/bluemix-letsencrypt/pull/31) had been bugging me. The failure induced by [`ibmcloud --quiet`](https://github.com/ibmjstart/bluemix-letsencrypt/issues/32) triggered me to look for the exit code of the `cf ssh` command. Not the code keeps trying in a loop, but without an _initial_ sleep.

The sleep seems to be required in about 50% if the cases I tried.